About app
Friends and Family Locator app helps you locate your loved ones on the google maps. It supports various features like "Create Group", "View Groups", "Plan a Meeting", "View Meetings", "My Invitations", "Mark Danger Location", "View Danger Locations" and "Panic Alert". It also supports OTP login using firebase. First you create a group by selecting contacts that are having account on this app and are in your contacts list. After an invite is sent to each member, if they accept the invite then they are added in the group otherwise not. Then you can select that particular group and locate those group members' location. You can also plan a meeting for a particular group and give a title, date, time and location to it. It also supports a chat system where group members can reply in the meeting whether they can come or not. One can mark danger location which is visible to all other users publically and also view and delete his marked danger location. Also when a user enters within a 100 meters range in any of the marked danger location he will get a notification that he is in a danger location. And lastly it supports a panic alert with which you can send a text message to all the group members whenever you are in a danger situation.